refactor: product 예외 처리 로직

+++ src/main/java/gift/product/ProductService.java
@@ -1,7 +1,10 @@
 package gift.product;
 
+import gift.exception.InvalidProduct;
 import java.util.List;
 import java.util.Optional;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -19,10 +22,11 @@ public class ProductService {
     }
 
     // 아이디로 상품이 존재하는지 확인이 아니고 상품 조회
-    public ProductResponseDto getProductById(Long id) {
-        Optional<Product> product = productRepository.findById(id);
-        Product pd = product.get();
-        return new ProductResponseDto(pd.getId(), pd.getName(), pd.getPrice(), pd.getImageUrl());
+    public Optional<Product> getProductById(Long id) {
+//        Optional<Product> product = productRepository.findById(id);
+//        Product pd = product.get();
+//        return new ProductResponseDto(pd.getId(), pd.getName(), pd.getPrice(), pd.getImageUrl());
+        return productRepository.findById(id);
     }
 
     // 상품 인서트
@@ -33,16 +37,27 @@ public class ProductService {
     // 상품 정보 수정
     public Product putProduct(Long id, ProductRequestDto productRequestDto) {
         Optional<Product> product = productRepository.findById(id);
+        if (product.isPresent()) {
+            Product pd = product.get();
+            pd.update(productRequestDto.name(), productRequestDto.price(), productRequestDto.url());
 
-        Product pd = product.get();
-        pd.update(productRequestDto.name(), productRequestDto.price(), productRequestDto.url());
+            return pd;
+        } else {
+            throw new InvalidProduct("유효하지 않은 상품입니다");
+        }
 
-        return pd;
+
     }
 
     // 상품 삭제
-    public void deleteProductById(Long id) {
-        productRepository.deleteById(id);
+    public HttpEntity<String> deleteProductById(Long id) {
+        Optional<Product> product = productRepository.findById(id);
+        if (product.isEmpty()) {
+            throw new InvalidProduct("유효하지 않은 상품입니다");
+        } else {
+            productRepository.deleteById(id);
+        }
+        return ResponseEntity.ok("성공적으로 삭제되었습니다");
     }
 
 }
